# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "contact" : { },
    "description" : "API for Acontext.",
    "title" : "Acontext API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "/api/v1"
  } ],
  "paths" : {
...
    "/session/{session_id}/events" : {
      "get" : {
        "description" : "Get events for a session with cursor-based pagination.",
        "parameters" : [ {
          "description" : "Session ID",
          "in" : "path",
          "name" : "session_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        }, {
          "description" : "Limit of events to return, default 50. Max 200.",
          "in" : "query",
          "name" : "limit",
          "schema" : {
            "type" : "integer"
          }
        }, {
          "description" : "Cursor for pagination.",
          "in" : "query",
          "name" : "cursor",
          "schema" : {
            "type" : "string"
          }
        }, {
          "description" : "Order by created_at descending if true, ascending if false (default false)",
          "in" : "query",
          "name" : "time_desc",
          "schema" : {
            "type" : "boolean"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_session__session_id__events_get_200_response"
                }
              }
            },
            "description" : "OK"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Get events for session",
        "tags" : [ "session" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Get events for a session\nevents = client.sessions.get_events(session_id, limit=50)\nfor event in events.items:\n    print(f\"{event.type}: {event.data}\")\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Get events for a session\nconst events = await client.sessions.getEvents(sessionId, { limit: 50 });\nfor (const event of events.items) {\n  console.log(`${event.type}: ${JSON.stringify(event.data)}`);\n}\n"
        } ]
      },
      "post" : {
        "description" : "Add a structured event to a session. Events are stored alongside messages and can be retrieved chronologically.",
        "parameters" : [ {
          "description" : "Session ID",
          "in" : "path",
          "name" : "session_id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/handler.AddEventReq"
              }
            }
          },
          "description" : "AddEvent payload",
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/_session__session_id__events_post_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "Invalid request"
          },
          "404" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/serializer.Response"
                }
              }
            },
            "description" : "Session not found"
          }
        },
        "security" : [ {
          "BearerAuth" : [ ]
        } ],
        "summary" : "Add event to session",
        "tags" : [ "session" ],
        "x-code-samples" : [ {
          "label" : "Python",
          "lang" : "python",
          "source" : "from acontext import AcontextClient\nfrom acontext.event import DiskEvent, TextEvent\n\nclient = AcontextClient(api_key='sk_project_token')\n\n# Add a disk event\nclient.sessions.add_event(session_id, DiskEvent(disk_id='xxxx', path='/data/report.csv', note='Uploaded report'))\n\n# Add a text event\nclient.sessions.add_event(session_id, TextEvent(text='User switched to dark mode'))\n"
        }, {
          "label" : "JavaScript",
          "lang" : "javascript",
          "source" : "import { AcontextClient, DiskEvent, TextEvent } from '@acontext/acontext';\n\nconst client = new AcontextClient({ apiKey: 'sk_project_token' });\n\n// Add a disk event\nawait client.sessions.addEvent(sessionId, new DiskEvent({ diskId: 'xxxx', path: '/data/report.csv', note: 'Uploaded report' }));\n\n// Add a text event\nawait client.sessions.addEvent(sessionId, new TextEvent({ text: 'User switched to dark mode' }));\n"
        } ],
        "x-codegen-request-body-name" : "payload"
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "handler.AddEventReq" : {
        "type" : "object"
      },
...
      "model.SessionEvent" : {
        "properties" : {
          "created_at" : {
            "type" : "string"
          },
          "data" : {
            "properties" : { },
            "type" : "object"
          },
          "id" : {
            "type" : "string"
          },
          "project_id" : {
            "type" : "string"
          },
          "session_id" : {
            "type" : "string"
          },
          "type" : {
            "type" : "string"
          },
          "updated_at" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "serializer.Response" : {
        "properties" : {
          "code" : {
            "type" : "integer"
          },
          "error" : {
            "type" : "string"
          },
          "msg" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "service.ListEventsOutput" : {
        "properties" : {
          "has_more" : {
            "type" : "boolean"
          },
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/model.SessionEvent"
            },
            "type" : "array"
          },
          "next_cursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "_session__session_id__events_get_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/service.ListEventsOutput"
            }
          },
          "type" : "object"
        } ]
      },
      "_session__session_id__events_post_200_response" : {
        "allOf" : [ {
          "$ref" : "#/components/schemas/serializer.Response"
        }, {
          "properties" : {
            "data" : {
              "$ref" : "#/components/schemas/model.SessionEvent"
            }
          },
          "type" : "object"
        } ]
      },
...
    },
    "securitySchemes" : {
      "BearerAuth" : {
        "description" : "Project Bearer token (e.g., \"Bearer sk-ac-xxxx\")",
        "in" : "header",
        "name" : "Authorization",
        "type" : "apiKey"
      }
    }
  },
...
}